back to ic2

+++ src/main/java/com/anet/qtr4tdm/common/entities/LaserTurretEntity.java
@@ -42,7 +42,7 @@ public class LaserTurretEntity extends TurretEntity {
     }
 
     @Override
-    public long getMaxEnergy() {
+    public double getMaxEnergy() {
         return 10000;
     }
